% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommKern.R
\docType{package}
\name{CommKern}
\alias{CommKern}
\title{CommKern}
\description{
The CommKern package provides a streamlined implementation in the design and
analysis of network community structures with specific applications to
neuroimaging data. The hierarchical multimodal spinglass (HMS) algorithm has
been developed as a novel community detection algorithm, while the
semiparametric kernel machine methods allow for statistical inference to be
performed to test for potential associations between these community
structures and an outcome of interest, whether binary or continuous.

This package was part of Alexandria Jensen's Ph.D. dissertation which was
overseen by her advisor Debashis Ghosh. Peter DeWitt provided extensive
mentorship on the creation of this package.
}
