% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_allegiance.R
\name{community_allegiance}
\alias{community_allegiance}
\title{Community allegiance}
\usage{
community_allegiance(comm_matrix)
}
\arguments{
\item{comm_matrix}{a matrix whose first column is the node label/id and all subsequent columns are different partitions}
}
\value{
alleg_matrix_norm, a matrix whose values are bounded in [0,1], where higher values indicate that a node belongs to a
single community over a higher proporition of runs
}
\description{
Description of the community allegiance function.
}
\details{
This function calculates the community allegiance of each node in a network. For node i, the stability of its 
allegiance to community A is calculated as the number of times where node i belongs to community A, divided by
the total number of runs. This measure is bounded in [0,1], where higher values of stability indicate
that a node belong to a single community across a greater number of runs.

The function returns the consensus partition, which is determined by the maximum average pairwise similarity.
}
