% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pairs.R
\name{count_pairs}
\alias{count_pairs}
\title{Count pairs}
\usage{
count_pairs(a, b, order)
}
\arguments{
\item{a}{a vector of classifications}

\item{b}{a vector of classifications}

\item{order}{a vector of permutations (coming from the order() function in base R)}
}
\value{
A list of five different vectors:
\itemize{
  \item{pair_nb}{a vector containing counts of nodes within all possible
  classification pairs from partitions a and b}
  \item{pair_a}{a vector of the same length as pair_nb, specifying the order
  of classifications in pair_nb from parition a}
  \item{pair_b}{a vector of the same length as pair_nb, specifying the order
  of classifications in pair_nb from parition b}
  \item{a_nb}{a vector containing counts of nodes within each class for
  partition a}
  \item{b_nb}{a vector containing counts of nodes within each class for
  partition b}
}
}
\description{
Description of the count pairs function.
}
\details{
A function to count pairs of integers or factors and identify the pair counts
}
