% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{SBM_net}
\title{Example Data Set(s) for Hierarchical Spinglass}
\format{
An object of class \code{list} of length 5.
}
\usage{
SBM_net
}
\description{
Example Data Set(s) for Hierarchical Spinglass
}
\keyword{datasets}
