% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{Node degree calculation}
\usage{
degree(adj_matrix_func, adj_matrix_str, vertex_df)
}
\arguments{
\item{adj_matrix_func}{the adjacency matrix for functional connectivity}

\item{adj_matrix_str}{the adjacency matrix for structural connectivity}

\item{vertex_df}{a data frame of node (or vertex) information}
}
\value{
a data frame to be incorporated into the network object
}
\description{
Description of the node degree calculation function.
}
\details{
This is an ancillary function that calculates the functional and structural degree of each network
node using the functional and structural adjacency matrices, respectively.
}
