% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\usage{
entropy(a, b)
}
\arguments{
\item{a}{a vector of classifications; this must be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{b}{a vector of classifications}
}
\value{
res, a list of four objects, including the two conditional entropies, the joint entropy, and the output
of sort_pairs
}
\description{
Description of the entropy function.
}
\details{
A function to compute the empirical entropy for two vectors of classifications and the joint entropy

The function returns a list of four different objects:
uv the joint entropy
u the conditional entropy of partition a
v the conditional entropy of partition b
sortPairs the output from the sort_pairs function ((see \code{\link{sort_pairs}} for more details))
}
