% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_distance.R
\name{ext_distance}
\alias{ext_distance}
\title{Extrinsic evaluation distance matrix creation}
\usage{
ext_distance(comm_df, variant = c("NMI", "Adj_RI", "purity"))
}
\arguments{
\item{comm_df}{a dataframe whose columns are different partitions. All partitions must have the same
set of nodes in order for this function to work and this dataframe should exclude a node ID column
for ease of computation.}

\item{variant}{a string in ('NMI', 'Adj_RI', 'purity') that calculates different external cluster
evaluation metrics.}
}
\value{
comm_dist, the extrinsic evaluation distance matrix to be used as input for the kernel function
}
\description{
Description of the extrinisic evaluation distance matrix creation function.
}
\details{
This function creates a distance matrix using the community output values from the hierarchical
multimodal spinglass algorithm. Because extrinsic evaluation metrics for clustering algorithms use
the underlying idea of similarity, distance is calculated as (1-similarity). The use of distance ensures
that the distance matrix will be positive and semi-definite, a requirement for its use in the kernel function.

The function returns an m x m matrix (where m is the number of partitions) to be used as input for
the kernel function.
}
