% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_start_temp.R
\name{find_start_temp}
\alias{find_start_temp}
\title{Starting temperature}
\usage{
find_start_temp(net, mod_matrix, spins, alpha, ts)
}
\arguments{
\item{net}{a \code{spinglass_net} object}

\item{mod_matrix}{mod_matrix}

\item{spins}{spins}

\item{alpha}{a double parameter balancing the use of the guidance matrix in modularity calculation}

\item{ts}{the starting temperature for the search, set to 1 within the algorithm}
}
\value{
the starting temperature that meets the criteria specified above
}
\description{
Description of the starting temperature function.
}
\details{
Within the spinglass algorithm, we would like to start from a temperature
with at least 95 of all proposed spin changes accepted in 50 sweeps over the network.
The function returns the temperature found.
}
