% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_network_perturb.R
\name{group_network_perturb}
\alias{group_network_perturb}
\title{Simulated group networks}
\usage{
group_network_perturb(
  n_nodes,
  n_comm,
  n_nets,
  perturb_prop,
  wcr,
  bcr,
  bfcr,
  fuzzy_comms
)
}
\arguments{
\item{n_nodes}{the number of nodes in each simulated network (will be the same across all networks)}

\item{n_comm}{the number of communities to be simulated in each network (will be the same across all networks)}

\item{n_nets}{the number of networks to simulate}

\item{perturb_prop}{the proportion of network nodes to randomly alter their community assignment within each network}

\item{wcr}{within community edge weights, sampled from a beta distribution; for example, c(8,8) will ask for the within
community edge weights to be sampled from a Beta(8,8) distribution}

\item{bcr}{between community edge weights, sampled from a beta distribution; for example, c(1,8) will ask for the between
community edge weights to be sampled from a Beta(1,8) distribution}

\item{bfcr}{fuzzy community edge weights, sampled from a beta distribution; for example, c(4,8) will ask for the
fuzzy community edge weights to be sampled from a Beta (4,8) distribution}

\item{fuzzy_comms}{the communities for which their distinction is 'fuzzy,' or not as distinct; fuzzy communities tend to have
higher between community edge weights; for example, c("comm_a","comm_c") will create a fuzzy distinction between
communities a and c}
}
\value{
net_perturb_list, a list of network dataframes containing nodes, their community assignment, node dyads, and edge weights
}
\description{
Description of the simulated group networks function.
}
\details{
This function creates a list of simulated networks, of which each network is in a dataframe format,
which describes describes the community assignment for each node in the network, and simulates the edge weights
based on whether the node dyad is: (a) withiin the same community; (b) between different communities, or (c)
between different communities, but designated as 'fuzzy' in their distinction from one another.

The function returns a list of dataframes detailing the nodes, node dyads, community assignments,
and edge weights for all dyads in each simulated network
}
