% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_mult_spin.R
\name{hms}
\alias{hms}
\title{Hierarchical multimodal spinglass algorithm}
\usage{
hms(input_net, spins, alpha, coolfact, false_pos, max_layers, parallel = FALSE)
}
\arguments{
\item{input_net}{a network object in list form (see the matrix_to_df() function for more details)}

\item{spins}{an integer indicating the maximum number of spins, or communities, that can be used}

\item{alpha}{a double parameter balancing the use of the guidance matrix in modularity calculation}

\item{coolfact}{a double parameter that indicates how quickly (or slowly) to cool the heathbath algorithm, typically set to be 0.95-0.99}

\item{false_pos}{a double parameter that indicates the level of false positives to allow the system to make (if ground truth is known), typically set to 0.01-0.05}

\item{max_layers}{an integer parameter that specifies the maximum number of layers of communities within the network}

\item{parallel}{a boolean operator (not currently used)}
}
\value{
comm_layers_tree a dataframe consisting of nodes and their community assignments across the layers
}
\description{
Description of the hierarchcial multimodal spinglass algorithm function.
}
\details{
This is the main function of the algorithm. After running checks on the input parameters, the algorithm
begins on the first layer of the network, finding the optimal configuration of nodes to communities using
the heatbath algorithm. Once the community assignments have been finalized, the set of nodes within each
of these communities is broken up and become their own subnetworks, on which the algorithm is applied again
to get further subnetwork community assignments. This continues until the maximum number of layers is reached.

The function returns a dataframe of nodes to communities across the layers.
}
