% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{init_config}
\alias{init_config}
\title{Initial configuration assignment}
\usage{
init_config(spin)
}
\arguments{
\item{spin}{an integer indicating the maximum number of spins, or communities, that can be used}
}
\value{
net list object with the communities portion of the list initially assigned
}
\description{
Description of the initial configuration assignment function.
}
\details{
Assigning an initial random configuration of spins to nodes if called with a negative 
argument or the spin used as argument when called with a positive one.
The function returns the initial random configuration of nodes to communities.
}
