% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel}
\alias{kernel}
\title{Distance-based kernel}
\usage{
kernel(mat, rho)
}
\arguments{
\item{mat}{a distance-based matrix}

\item{rho}{a bandwidth/scaling parameter whose optimal value is solved for
within the larger score function}
}
\value{
the value of the kernel
}
\description{
Description of the distance-based kernel function
}
\details{
This is an ancillary function that is passed into the score function, used
for calculating the distance-based kernel.

The function returns an m x m matrix (where m is the number of networks) to
be used as input for the kernel function.
}
