% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_lin_semiparm.R
\name{score_cont_semiparam}
\alias{score_cont_semiparam}
\title{Semiparametric score function for distance-based kernel and continuous outcome.}
\usage{
score_cont_semiparam(outcome, covars, dist_mat, grid_gran = 5000)
}
\arguments{
\item{outcome}{a numeric vector containing the continuous outcome variable (in the same ID order as dist_mat)}

\item{covars}{a dataframe containing the covariates to be modeled parametrically (should NOT include an ID variable)}

\item{dist_mat}{a square distance matrix}

\item{grid_gran}{a numeric value specifying the grid search length, preset to 5000}
}
\value{
the score function p-value
}
\description{
Description of the semiparametric score function for distance-based kernel function and continuous outcome.
}
\details{
This is the main function that calculates the p-value associated with a semiparametric kernel test
of association between the kernel and continuous outcome variable. A null model (where the kernel is not
associated with the outcome) is initially fit. Then, the variance of Y_{i}|X_{i} is used as the basis for the
score test, S\left(\rho\right) = \dfrac{Q_{\tau}\left(\hat{\beta_0},\rho\right)-\mu_Q}{\sigma_Q}. However,
because \rho disappears under the null hypothesis, we run a grid search over a range of values of \rho (the bounds
of which were derived by Liu et al. in 2008). This grid search gets the upper bound for the score test's p-value.
This function is tailored for the underlying model y_{i} = x_{i}^{T}\beta + h\left(z_{i}\right) + e_{i}, where h\left(\cdot\right) is
the kernel function, z_{i} is a multidimensional array of variables, x_{i} is a vector or matrix of covariates, \beta is a vector
of regression coefficients, and y_{i} is a continuous outcome taking values in the real numbers.

The function returns an numeric p-value for the kernel score test of association.
}
