% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simasd_array}
\alias{simasd_array}
\title{Simulated Array}
\format{
An object of class \code{array} of dimension 49 x 80 x 80.
}
\usage{
simasd_array
}
\description{
Simulated Array
}
\keyword{datasets}
