% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simasd_comm_df}
\alias{simasd_comm_df}
\title{Simulated partitions of nodes to communities from HMS algorithm}
\format{
A dataframe with 80 rows and 49 columns, where rows correspond to
nodes within the simulated networks and columns correspond to the subject ID.
}
\usage{
simasd_comm_df
}
\description{
A dataset of partitions of nodes to communities from simulated group-level
networks with community structures. This dataset is complementary to the
simasd_covars dataset, which contains the demographic information related
to this dataset. For more information on how these group-level networks 
were simulated, please refer to the example script titled "beta_simulation_data.set.R".
The variables are as follows:
}
\keyword{datasets}
