% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simasd_hamil_df}
\alias{simasd_hamil_df}
\title{Simulated Hamiltonian values from HMS algorithm}
\format{
A dataframe with 49 rows and 2 columns:
\describe{
  \item{id}{a generic ID, corresponding to the id variable in simasd_covars}
  \item{hamil}{Hamiltonian value calculated from running the simulated network through the
    HMS algorithm, a numeric value}
}
}
\usage{
simasd_hamil_df
}
\description{
A dataset of Hamiltonian values from simulated group-level networks with
community structure. This dataset is complementary to the simasd_covars
dataset, which contains the demographic information related to this dataset.
For more information on how these group-level networks were simulated, please
refer to the example script titled "beta_simulation_data.set.R".
The variables are as follows:
}
\keyword{datasets}
