% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_pairs.R
\name{sort_pairs}
\alias{sort_pairs}
\title{Sort pairs}
\usage{
sort_pairs(a, b)
}
\arguments{
\item{a}{a vector of classifications}

\item{b}{a vector of classifications}
}
\value{
res a list of six objects used asthe basis to calculate many cluster evaluation metrics,
like NMI, ARI, and the Rand z-score
}
\description{
Description of the sort pairs function.
}
\details{
A function to sort pairs of integers or factors and identify the pairs

The function returns a list of six different objects:
levels (a list of the classes within each of the partitions a and b)
n_ij (a vector containing counts of nodes within all possible classification pairs from partitions a and b)
n_i. (a vector of the same length as pair_nb, specifying the order of classifications in pair_nb from parition a)
n_.j (a vector of the same length as pair_nb, specifying the order of classifications in pair_nb from parition b)
pair_a (a vector containing counts of nodes within each class for partition a)
pair_b (a vector containing counts of nodes within each class for partition b)
}
