% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_matrix_to_df.R
\name{subset_matrix_to_df}
\alias{subset_matrix_to_df}
\title{Convert matrices to dataframe list for subnetworks}
\usage{
subset_matrix_to_df(func_matrix, str_matrix)
}
\arguments{
\item{func_matrix}{a network object in list form (see the matrix_to_df() function for more details)}

\item{str_matrix}{an integer indicating the maximum number of spins, or communities, that can be used}
}
\value{
a list of dataframes for the subnetwork
}
\description{
Description of the convert matrices to dataframe list for subnetworks function.
}
\details{
This is an ancillary function that creates a dataframe list for the subnetworks created using
the multimodal hierarchical spinglass algorithm.
 
The function returns a dataframe list containing the functional matrix, structural matrix, a dataframe
of the functional edge weights, a dataframe of the structural edge weights, and nodal information (functional
degree, structural degree, community assignment, and label information)
}
