% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tr}
\alias{tr}
\title{Trace}
\usage{
tr(x, ...)
}
\arguments{
\item{x}{a square matrix}

\item{...}{arguments passed to \code{\link{sum}}}
}
\value{
The trace, the sum of the diagonal elements, of the square matrix
\code{x}
}
\description{
Return the trace of a square matrix
}
