% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up_low.R
\name{up_low}
\alias{up_low}
\title{Bounds of grid search function}
\usage{
up_low(dist_mat)
}
\arguments{
\item{dist_mat}{a square distance matrix}
}
\value{
a square matrix of the same dimensions of the input matrix, comprised
of the sum square differences.
}
\description{
Description of the bounds of grid search function.
}
\details{
This ancillary function finds the upper and lower bounds of the grid search
implemented in the kernel score test.

The function returns an m x m matrix (where m is the number of networks) to
be used as input for the kernel function.
}
