% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zrand.R
\name{zrand}
\alias{zrand}
\title{Rand z-score}
\usage{
zrand(part1, part2)
}
\arguments{
\item{part1}{a partition of nodes to communities or clusters}

\item{part2}{a partition of nodes to communities or clusters}
}
\value{
zRand, the Rand z-score between two paritions
}
\description{
Description of the Rand z-score function.
}
\details{
This is an ancillary function that calculates the Rand z-score between two partitions, which is
used in the consensus simularity function

The function returns a z-score, which is used in the calculation of the consensus paritition
}
