% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_plot.R
\name{community_plot}
\alias{community_plot}
\title{Communities by layer plot}
\usage{
community_plot(comm_data, node_data)
}
\arguments{
\item{comm_data}{is a long dataframe with three columns: node_id, layer, and comm; this can be created from the comm_layers_tree object using the tidyr::gather() function}

\item{node_data}{is a melted version of the network adjacency matrix, using the reshape2::melt() function}
}
\value{
A 1x2 panel plot.

the vertex_df dataframe to be incorporated into the network object
}
\description{
Description of the communities by layer plot function.
}
\details{
This is an ancillary function that creates the plots seen in the manuscript, with a heatmap-style
plot on the top panel, derived from a network adjacency matrix, and a community assignment plot
on the bottom panel, separated by layer.
}
\examples{

data(SBM_net)
SBM_netcomm <- hms(input_net=SBM_net,spins=4,alpha=0,coolfact=0.90,false_pos=0.05,max_layers=3)

community_plot(SBM_netcomm$comm_layers_tree,SBM_netcomm$func_matrix)

}
