% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{SBM_net}
\alias{simasd_covars}
\alias{simasd_hamil_df}
\title{Example Data Set(s) for Hierarchical Spinglass}
\format{
An object of class \code{spinglass_net} of length 5.

An object of class \code{data.frame} with 49 rows and 8 columns.

An object of class \code{data.frame} with 49 rows and 2 columns.
}
\usage{
SBM_net

simasd_covars

simasd_hamil_df
}
\description{
Example Data Set(s) for Hierarchical Spinglass
}
\keyword{datasets}
