% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\usage{
entropy(a, b)
}
\arguments{
\item{a}{a vector of classifications; this must be a vector of characters,
integers, numerics, or a factor, but not a list.}

\item{b}{a vector of classifications}
}
\value{
a list of four objects:
\itemize{
  \item{\code{uv}}{ the joint entropy}
  \item{\code{u}}{ the conditional entropy of partition \code{a}}
  \item{\code{v}}{ the conditional entropy of partition \code{b}}
  \item{\code{sort_pairs}}{ the output from the sort_pairs function}
}
}
\description{
Description of the entropy function.
}
\details{
A function to compute the empirical entropy for two vectors of classifications and the joint entropy
}
\seealso{
\code{\link{sort_pairs}}
}
