% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{Example Function}
\usage{
example_function(x, ymat, ...)
}
\arguments{
\item{x}{a numeric value}

\item{ymat}{a matrix}

\item{...}{not currentl used}
}
\value{
the scalar product of x by ymat
}
\description{
Description of the Example Function.
}
\details{
Details....  (can be many paragraphs)
}
