% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_adj_perturb.R
\name{group_adj_perturb}
\alias{group_adj_perturb}
\title{Group adjacency matrices}
\usage{
group_adj_perturb(group_network_list, n_nets, n_nodes)
}
\arguments{
\item{group_network_list}{the output from \code{\link{group_network_perturb}}, which is a list of dataframes
detailing nodes, community assignments of each node, and edge weights between each dyad of nodes}

\item{n_nets}{the number of networks simulated}

\item{n_nodes}{the number of nodes in each simulated network (will be the same across all networks)}
}
\value{
adj_array, an array of adjacency matrices of dimension (n_nets x n_nodes x n_nodes)
}
\description{
Description of the simulated group adjacency matrices function.
}
\details{

}
\seealso{
\code{\link{group_network_perturb}}
}
