% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_pairs.R
\name{sort_pairs}
\alias{sort_pairs}
\title{Sort pairs}
\usage{
sort_pairs(a, b)
}
\arguments{
\item{a}{a vector of classifications}

\item{b}{a vector of classifications}
}
\value{
a list of six objects used as the basis to calculate many cluster evaluation metrics,
like NMI, ARI, and the Rand z-score.
\itemize{
  \item{levels}{a list of the classes within each of the partitions a and b}
  \item{n_ij}{a vector containing counts of nodes within all possible classification pairs from partitions a and b}
  \item{n_i.}{a vector of the same length as pair_nb, specifying the order of classifications in pair_nb from partition a}
  \item{n_.j}{a vector of the same length as pair_nb, specifying the order of classifications in pair_nb from partition b}
  \item{pair_a}{a vector containing counts of nodes within each class for partition a}
  \item{pair_b}{a vector containing counts of nodes within each class for partition b}
}
}
\description{
Description of the sort pairs function.
}
\details{
A function to sort pairs of integers or factors and identify the pairs
}
